# Add same_shape case for BiasDropout