--- a/onnxruntime/core/graph/contrib_ops/contrib_defs.cc
+++ b/onnxruntime/core/graph/contrib_ops/contrib_defs.cc
@@ -2863,7 +2863,7 @@ It's an extension of Gelu. It takes the sum of input A and bias input B as the i
       .Attr("seed", "(Optional) Seed to the random generator, if not specified we will auto generate one.", AttributeProto::INT, OPTIONAL_VALUE)
       .AllowUncheckedAttributes()
       .Input(0, "data", "The input data as Tensor.", "T")
-      .Input(1, "bias", "The bias input, a vector with the same shape as last dim of data", "T")
+      .Input(1, "bias", "The bias input, a vector with the same shape as last dim of data OR same shape with data", "T")
       .Input(2, "residual", "The residual input, must have the same shape as data", "T", OpSchema::Optional)
       .Input(3, "ratio",
              "The ratio of random dropout, with value in [0, 1). If this input was not set, "