… onnxruntime/test/testdata/transform/fusion/bias_dropout_residual_same_shape_fusion.onnx
… onnxruntime/test/testdata/transform/fusion/bias_dropout_same_shape_fusion.onnx