**Description**: Add same_shape case for BiasDropout

**Motivation and Context**
Currently `BiasDropout` requires bias to be a 1-D tensor, this PR adds a new case where bias has the same shape as the data.

In Megatron-LM GPT-2 model, some perf gains are obtained:
Before: Add + Dropout + Add, 120.632 μs/layer * 24 layers, 44.259 ms (forward) + 52.696 ms (backward)
After: BiasDropout, 75.8 μs/layer * 24 layers, 43.316 ms (forward) + 52.226 (backward)

kernel saving: (120.632 - 75.8) μs * 24 = 1076 μs
real time saving for fw+bw: 1.46 %