@SherlockNoMad 2021-09-28
Ethan just introduced a vectorized version of dropout kernels, need to resolve conflicts.