@mindest 2021-09-28
@SherlockNoMad just notice this, do we need to restrict bias' dim to be a value? Could it be a param? https://github.com/microsoft/onnxruntime/blob/d3f859fe3063bdfc4b2fdadbd46985b3742b9aec/onnxruntime/core/optimizer/bias_dropout_fusion.cc#L112-L114