@pengwa 2021-09-29
> @SherlockNoMad just notice this, do we need to restrict bias' dim to be a value? Could it be a param?
> 
> https://github.com/microsoft/onnxruntime/blob/d3f859fe3063bdfc4b2fdadbd46985b3742b9aec/onnxruntime/core/optimizer/bias_dropout_fusion.cc#L112-L114

I think the question is, is there any reason we do the restriction (the bias' dim must be a concreate value, cannot be a symbolic dim) when you implement the kernels at that time?  @SherlockNoMad