@SherlockNoMad 2021-10-11
> e this, do we need to restrict bias' dim to be a value? Could it be a param

yes, it can also be a param. and this case should also be considered. 
e.g. two shape of [batch, seq, 128], [batch, seq, 128] should be considered identical.